refacto(read): move tokens code to new module `_tokens`
