+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 (internal)
 - `_parsing.py` renamed to `_reading.py`
+- create `_tokens.py`
 
 ### fixed
